travis: use custom swdb.lst
Verifying the downloaded one with gpgv seems to fail.
No idea why... or rather it cannot find the public keys
even if i add them to the default gpg1 and gpg2 keyrings
plus ~/.gnupg/trustedkeys.gpg

I doubt this will work on its own because i think the
existance of swdb.lst on my local machine is an artifact
of previous attempts to build. Trying this to figure that
out.