wip: first step towards recommendations
For now mainly some glue code in Account and a new class Recommendation.
Idea is to have the recommendation logic encapsulated in Recommendation
as a value object (immutable state and no sideeffects). That way we can
easily test a bunch of corner cases in isolation while keeping the code
in Account at a minimum.