fix doc build
Just fixing the error messages of tox -e doc.

This is probably not what we want in the end. For example the
enumerated lists at the end of oob_verification will be
interpreted as independent lists because of the empty lines
separating them i think. But that will only become interesting
later. For now this hopefully fixes the build.