encrypt with bot
Note that in order to properly decrypt the message we first
need to process the AC Header so we have the key to verify
the signature.

If we leave out this step it will lead to a response code of
2 from gnupg and raise an unprintable InvocationFailure:
```
        if ret != 0 or (strict and err):
            raise self.InvocationFailure(ret, " ".join(args),
>                                        out=out, err=err)
E           muacrypt.bingpg.InvocationFailure: <unprintable InvocationFailure object>

.tox/py35/lib/python3.5/site-packages/muacrypt/bingpg.py:190
```

I will see if i can turn this into something printable as i believe
we might run into it more often.